requirements.txt
fastapi[standard]==0.115.0
aiogram==3.22.0
sqlmodel==0.0.30
python-dotenv==1.2.1